Extract Elapsed into module
